+++ src/main/java/pl/Ljimmex/fractionCore/module/ModuleCommand.java
@@ -1,7 +1,6 @@
 package pl.Ljimmex.fractionCore.module;
 
 import org.bukkit.command.Command;
-import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.jetbrains.annotations.NotNull;